Begin replacing spec classes with Pydantic models
