+++ .pre-commit-config.yaml
@@ -20,7 +20,7 @@ repos:
 -   repo: https://github.com/astral-sh/ruff-pre-commit
     rev: v0.13.2
     hooks:
-    -   id: ruff
+    -   id: ruff-check
 # -   repo: https://github.com/econchick/interrogate
 #     rev: 1.5.0
 #     hooks:
