+++ src/hdmf/spec/__init__.py
@@ -1,5 +1,9 @@
 from .catalog import SpecCatalog
 from .namespace import NamespaceCatalog, SpecNamespace, SpecReader
-from .spec import (AttributeSpec, DatasetSpec, DtypeHelper, DtypeSpec, GroupSpec, LinkSpec,
-                   NAME_WILDCARD, RefSpec, Spec)
+# from .spec import (AttributeSpec, DatasetSpec, DtypeHelper, DtypeSpec, GroupSpec, LinkSpec,
+#                    NAME_WILDCARD, RefSpec, Spec)
+from .spec2 import (
+    AttributeSpec, DatasetSpec, DtypeHelper, DtypeSpec, GroupSpec, LinkSpec, RefSpec, Spec, BaseStorageSpec,
+    QuantityEnum
+)
 from .write import NamespaceBuilder, SpecWriter, export_spec
