+++ tests/unit/spec_tests/test_dtype_spec.py
@@ -3,8 +3,6 @@ from hdmf.testing import TestCase
 
 
 class DtypeSpecHelper(TestCase):
-    def setUp(self):
-        pass
 
     def test_recommended_dtypes(self):
         self.assertListEqual(DtypeHelper.recommended_primary_dtypes,
@@ -16,53 +14,53 @@ class DtypeSpecHelper(TestCase):
         self.assertSetEqual(a, DtypeHelper.valid_primary_dtypes)
 
     def test_simplify_cpd_type(self):
-        compound_type = [DtypeSpec('test', 'test field', 'float'),
-                         DtypeSpec('test2', 'test field2', 'int')]
+        compound_type = [
+            DtypeSpec(name='test', doc='test field', dtype='float'),
+            DtypeSpec(name='test2', doc='test field2', dtype='int'),
+        ]
         expected_result = ['float', 'int']
         result = DtypeHelper.simplify_cpd_type(compound_type)
         self.assertListEqual(result, expected_result)
 
     def test_simplify_cpd_type_ref(self):
-        compound_type = [DtypeSpec('test', 'test field', 'float'),
-                         DtypeSpec('test2', 'test field2', RefSpec(target_type='MyType', reftype='object'))]
+        compound_type = [
+            DtypeSpec(name='test', doc='test field', dtype='float'),
+            DtypeSpec(name='test2', doc='test field2', dtype=RefSpec(target_type='MyType', reftype='object')),
+        ]
         expected_result = ['float', 'object']
         result = DtypeHelper.simplify_cpd_type(compound_type)
         self.assertListEqual(result, expected_result)
 
     def test_check_dtype_ok(self):
-        self.assertEqual('int', DtypeHelper.check_dtype('int'))
+        DtypeHelper.check_dtype(dtype='int')  # should not raise
 
     def test_check_dtype_bad(self):
-        msg = "dtype 'bad dtype' is not a valid primary data type."
-        with self.assertRaisesRegex(ValueError, msg):
-            DtypeHelper.check_dtype('bad dtype')
+        with self.assertRaisesRegex(ValueError, r"'bad dtype' is not a valid primary data type\."):
+            DtypeHelper.check_dtype(dtype='bad dtype')
 
     def test_check_dtype_ref(self):
         refspec = RefSpec(target_type='target', reftype='object')
-        self.assertIs(refspec, DtypeHelper.check_dtype(refspec))
+        DtypeHelper.check_dtype(dtype=refspec)  # should not raise
 
     def test_is_allowed(self):
-        self.assertTrue(DtypeHelper.is_allowed_dtype('int32', 'int'))
-        self.assertTrue(DtypeHelper.is_allowed_dtype('float64', 'float'))
-        self.assertFalse(DtypeHelper.is_allowed_dtype('int32', 'float'))
-        self.assertFalse(DtypeHelper.is_allowed_dtype('string', 'int'))
-        self.assertTrue(DtypeHelper.is_allowed_dtype('object', 'object'))
-        self.assertTrue(DtypeHelper.is_allowed_dtype('int64', 'numeric'))
-        self.assertTrue(DtypeHelper.is_allowed_dtype('float32', 'numeric'))
-        self.assertFalse(DtypeHelper.is_allowed_dtype('string', 'numeric'))
-        self.assertTrue(DtypeHelper.is_allowed_dtype('numeric', 'numeric'))
+        self.assertTrue(DtypeHelper.is_allowed_dtype(new='int32', orig='int'))
+        self.assertTrue(DtypeHelper.is_allowed_dtype(new='float64', orig='float'))
+        self.assertFalse(DtypeHelper.is_allowed_dtype(new='int32', orig='float'))
+        self.assertFalse(DtypeHelper.is_allowed_dtype(new='string', orig='int'))
+        self.assertTrue(DtypeHelper.is_allowed_dtype(new='object', orig='object'))
+        self.assertTrue(DtypeHelper.is_allowed_dtype(new='int64', orig='numeric'))
+        self.assertTrue(DtypeHelper.is_allowed_dtype(new='float32', orig='numeric'))
+        self.assertFalse(DtypeHelper.is_allowed_dtype(new='string', orig='numeric'))
+        self.assertTrue(DtypeHelper.is_allowed_dtype(new='numeric', orig='numeric'))
 
-        msg = "Unknown dtype 'bad dtype'"
-        with self.assertRaisesRegex(ValueError, msg):
-            DtypeHelper.is_allowed_dtype('int32', 'bad dtype')
+        with self.assertRaisesRegex(ValueError, r"Invalid dtype 'bad dtype'\."):
+            DtypeHelper.is_allowed_dtype(new='int32', orig='bad dtype')
 
 
 class DtypeSpecTests(TestCase):
-    def setUp(self):
-        pass
 
     def test_constructor(self):
-        spec = DtypeSpec('column1', 'an example column', 'int')
+        spec = DtypeSpec(name='column1', doc='an example column', dtype='int')
         self.assertEqual(spec.doc, 'an example column')
         self.assertEqual(spec.name, 'column1')
         self.assertEqual(spec.dtype, 'int')
@@ -73,23 +71,10 @@ class DtypeSpecTests(TestCase):
         self.assertEqual(spec.name, 'column1')
         self.assertEqual(spec.dtype, 'int')
 
-    def test_invalid_refspec_dict(self):
-        """Test missing or bad target key for RefSpec."""
-        msg = "'dtype' must have the key 'target_type'"
-        with self.assertRaisesWith(ValueError, msg):
-            DtypeSpec.assertValidDtype({'no target': 'test', 'reftype': 'object'})
-
     def test_refspec_dtype(self):
-        # just making sure this does not cause an error
-        DtypeSpec('column1', 'an example column', RefSpec('TimeSeries', 'object'))
+        # make sure this does not cause an error
+        DtypeSpec(name='column1', doc='an example column', dtype=RefSpec(target_type='TimeSeries', reftype='object'))
 
     def test_invalid_dtype(self):
-        msg = "dtype 'bad dtype' is not a valid primary data type."
-        with self.assertRaisesRegex(ValueError, msg):
-            DtypeSpec('column1', 'an example column', dtype='bad dtype')
-
-    def test_is_ref(self):
-        spec = DtypeSpec('column1', 'an example column', RefSpec('TimeSeries', 'object'))
-        self.assertTrue(DtypeSpec.is_ref(spec))
-        spec = DtypeSpec('column1', 'an example column', 'int')
-        self.assertFalse(DtypeSpec.is_ref(spec))
+        with self.assertRaisesRegex(ValueError, r"dtype 'bad dtype' is not a valid primary data type\."):
+            DtypeSpec(name='column1', doc='an example column', dtype='bad dtype')
