+++ tests/unit/spec_tests/test_ref_spec.py
@@ -1,4 +1,4 @@
-import json
+from pydantic import ValidationError
 
 from hdmf.spec import RefSpec
 from hdmf.testing import TestCase
@@ -7,11 +7,16 @@ from hdmf.testing import TestCase
 class RefSpecTests(TestCase):
 
     def test_constructor(self):
-        spec = RefSpec('TimeSeries', 'object')
+        spec = RefSpec(target_type='TimeSeries', reftype='object')
         self.assertEqual(spec.target_type, 'TimeSeries')
         self.assertEqual(spec.reftype, 'object')
-        json.dumps(spec)  # to ensure there are no circular links
+        spec_dict = spec.model_dump(exclude_unset=True)
+        expected = {
+            'target_type': 'TimeSeries',
+            'reftype': 'object',
+        }
+        self.assertDictEqual(spec_dict, expected)
 
     def test_wrong_reference_type(self):
-        with self.assertRaises(ValueError):
-            RefSpec('TimeSeries', 'unknownreftype')
+        with self.assertRaisesRegex(ValidationError, r"reftype\s*Input should be 'object'"):
+            RefSpec(target_type='TimeSeries', reftype='unknownreftype')
